+++ ui/src/assets.d.ts
@@ -7,3 +7,8 @@ declare module '*.webp' {
   const src: string;
   export default src;
 }
+
+declare module '*.ogg' {
+  const src: string;
+  export default src;
+}